Refactor: Integrate the functions of verifying problem into the newly created class Verifier, and rename namespace "presolve" with "preprocess".
