+++ cmake/test/CMakeLists.txt
@@ -38,8 +38,7 @@ link_directories(
 include(test_model_component.cmake)
 include(test_model.cmake)
 include(test_neighborhood.cmake)
-include(test_presolver.cmake)
-include(test_verifier.cmake)
+include(test_preprocess.cmake)
 include(test_utility.cmake)
 include(test_solver.cmake)
 include(test_solution.cmake)
