+++ cmake/test/test_preprocess.cmake
@@ -1,19 +1,19 @@
 
-file(GLOB TEST_MODEL_SOURCE ${TOP_DIR}/test/presolver/*.cpp)
+file(GLOB TEST_MODEL_SOURCE ${TOP_DIR}/test/preprocess/*.cpp)
 
 add_executable(
-    test_presolver
+    test_preprocess
     ${TEST_MODEL_SOURCE}
 )
 
 target_include_directories(
-    test_presolver
+    test_preprocess
     PUBLIC ${TOP_DIR}/printemps/
     PUBLIC ${TOP_DIR}/external/include/
 )
 
 target_link_libraries(
-    test_presolver
+    test_preprocess
     PUBLIC gtest
     PUBLIC gtest_main
     PUBLIC pthread
@@ -21,5 +21,5 @@ target_link_libraries(
 
 add_test(
     NAME test_presolver
-    COMMAND test_presolver --gtest_output=xml
+    COMMAND test_preprocess --gtest_output=xml
 )
