+++ printemps/preprocess/dependent_intermediate_variable_extractor.h
@@ -3,11 +3,11 @@
 // Released under the MIT license
 // https://opensource.org/licenses/mit-license.php
 /*****************************************************************************/
-#ifndef PRINTEMPS_PRESOLVER_DEPENDENT_INTERMEDIATE_VARIABLE_EXTRACTOR_H__
-#define PRINTEMPS_PRESOLVER_DEPENDENT_INTERMEDIATE_VARIABLE_EXTRACTOR_H__
+#ifndef PRINTEMPS_PREPROCESS_DEPENDENT_INTERMEDIATE_VARIABLE_EXTRACTOR_H__
+#define PRINTEMPS_PREPROCESS_DEPENDENT_INTERMEDIATE_VARIABLE_EXTRACTOR_H__
 
 namespace printemps {
-namespace presolver {
+namespace preprocess {
 /*****************************************************************************/
 template <class T_Variable, class T_Expression>
 class DependentIntermediateVariableExtractor {
@@ -216,7 +216,7 @@ class DependentIntermediateVariableExtractor {
     }
 };
 
-}  // namespace presolver
+}  // namespace preprocess
 }  // namespace printemps
 #endif
 /*****************************************************************************/
