+++ printemps/preprocess/flippable_variable_pair.h
@@ -3,11 +3,11 @@
 // Released under the MIT license
 // https://opensource.org/licenses/mit-license.php
 /*****************************************************************************/
-#ifndef PRINTEMPS_PRESOLVER_FLIPPABLE_VARIABLE_PAIR_H__
-#define PRINTEMPS_PRESOLVER_FLIPPABLE_VARIABLE_PAIR_H__
+#ifndef PRINTEMPS_PREPROCESS_FLIPPABLE_VARIABLE_PAIR_H__
+#define PRINTEMPS_PREPROCESS_FLIPPABLE_VARIABLE_PAIR_H__
 
 namespace printemps {
-namespace presolver {
+namespace preprocess {
 /*****************************************************************************/
 template <class T_Variable, class T_Expression>
 struct FlippableVariablePair {
@@ -45,7 +45,7 @@ struct FlippableVariablePair {
         this->overlap_rate             = 0;
     }
 };
-}  // namespace presolver
+}  // namespace preprocess
 }  // namespace printemps
 #endif
 /*****************************************************************************/
