+++ printemps/preprocess/selection_extractor.h
@@ -3,11 +3,11 @@
 // Released under the MIT license
 // https://opensource.org/licenses/mit-license.php
 /*****************************************************************************/
-#ifndef PRINTEMPS_PRESOLVER_SELECTION_EXTRACTOR_H__
-#define PRINTEMPS_PRESOLVER_SELECTION_EXTRACTOR_H__
+#ifndef PRINTEMPS_PREPROCESS_SELECTION_EXTRACTOR_H__
+#define PRINTEMPS_PREPROCESS_SELECTION_EXTRACTOR_H__
 
 namespace printemps {
-namespace presolver {
+namespace preprocess {
 /*****************************************************************************/
 template <class T_Variable, class T_Expression>
 class SelectionExtractor {
@@ -286,7 +286,7 @@ class SelectionExtractor {
         return m_selections;
     }
 };
-}  // namespace presolver
+}  // namespace preprocess
 }  // namespace printemps
 #endif
 /*****************************************************************************/
